refactor: move head to its own file
