+++ .vitepress/config/index.ts
@@ -4,12 +4,13 @@ import * as path from "node:path";
 import * as process from "node:process";
 import bytecode from "syntax-java-bytecode/java-bytecode.tmLanguage.json";
 import mcfunction from "syntax-mcfunction/mcfunction.tmLanguage.json";
+import { SiteConfig } from "vitepress";
 import { tabsMarkdownPlugin } from "vitepress-plugin-tabs";
 import defineVersionedConfig from "vitepress-versioning-plugin";
 import { transformFile, transformFilesPlugin } from "../plugins/transformFiles";
 import { Fabric } from "../types.d";
 import { getLocales } from "./i18n";
-import { transformHead, transformItems } from "./transform";
+import { transformHead } from "./head";
 
 const latestVersion = fs
   .readFileSync(
@@ -93,7 +94,13 @@ export default defineVersionedConfig(
 
     sitemap: {
       hostname,
-      transformItems,
+      transformItems: (items) => {
+        const config = (globalThis as any).VITEPRESS_CONFIG as SiteConfig;
+        return items.filter((i) => {
+          const relativePath = i.url.replace(hostname, "");
+          return !config.rewrites.inv[relativePath]?.startsWith("versions/");
+        });
+      },
     },
 
     srcExclude: ["README.md", "versions/1.21.10", ...(typeof env === "number" ? ["versions"] : [])],
